WIP: Add user modification to UserDb
One thing to note is that there seems to be a form wizard pattern trying
to burst out of this. I haven't done that yet, but I think a lot of the
control flow would be simplified with that in place.

I've also made the program print the user's phone number as part of
their records, because why else should we store it?